Add block validity prover implementation (#130)
* Rename block synchronizer

* Add types of witness

* Modify ValidityWitness

* wip

* Sync deposit tree

* Sync account tree

* small fix

* Implement validity function (wip)

* Fix Merkle tree

* Fix by linter

* Remove unnecessary file

* Connect block validity prover

* small fix

* Fix /transaction API

* Fix ticker

* Add private balance validity prover

* Modify private balance prover

* Modify private balance validity prover

* wip

* Revert "wip"

This reverts commit ec0503dfe7536efac870a47cfafe85f14072a177.

* Add database definition

* Add historical Merkle tree

* Fix Merkle tree

* wip

* wip

* wip

* Fix block validity prover

* Add deposits table

* wip

* wip

* wip

* Fix deposit_index column

* wip

* Fix prover

* Test using mock wallet and sample data (#126)

* Create sample data

* Pass e2e test

* Add request ID in ProofResponse

* wip

* Integrate withdrawal prover

* Work Gnark server

* Fix submitWithdrawalProof

* wip

* Add SyncLocally

* wip

* Separate client side code

* Remove SyncBlockProverWithBlockNumber from client code

* wip

* FIx block update

* Modify MockWallet

* wip

* Pass deposit proof

* wip

* Fix nonce type

* Remove BlockNumberByDepositIndex

* Fix FetchUpdateWitness

* Fix PostBlock

* Modify GetDepositInfoByHash

* Fix txDetails

* Remove PostBlock from BlockValidityProver

* Modify ValidityWitness

* Connect Data Store Vault

* (wip) FixLocally

* small fix

* Fix SncLocally

* Add deposit synchronizer

* Remove unnecessary interface

* Remove validityWitnesses

* Modify BlockAuxInfo

* Store block proof

* Scan block tree leaves

* Fix Prove function of Merkle tree

* Add LastPostedBlockNumber

* Fix block validity prover

* Fix user_address column

* Fix asset tree

* Add GetLatestBackupBalanceByUserAddress

* small fix

* Fix by linter

+++ internal/tree/account_tree.go
@@ -59,12 +59,12 @@ func (t *AccountTree) ProveMembership(publicKey *big.Int) (proof *IndexedMembers
 	return t.inner.ProveMembership(publicKey)
 }
 
-func (t *AccountTree) Insert(publicKey *big.Int, lastSentBlockNumber uint64) (*IndexedInsertionProof, error) {
-	return t.inner.Insert(publicKey, lastSentBlockNumber)
+func (t *AccountTree) Insert(publicKey *big.Int, lastSentBlockNumber uint32) (*IndexedInsertionProof, error) {
+	return t.inner.Insert(publicKey, uint64(lastSentBlockNumber))
 }
 
-func (t *AccountTree) Update(publicKey *big.Int, lastSentBlockNumber uint64) (*IndexedUpdateProof, error) {
-	return t.inner.Update(publicKey, lastSentBlockNumber)
+func (t *AccountTree) Update(publicKey *big.Int, lastSentBlockNumber uint32) (*IndexedUpdateProof, error) {
+	return t.inner.Update(publicKey, uint64(lastSentBlockNumber))
 }
 
 func NewDummyAccountRegistrationProof(height uint8) *IndexedInsertionProof {
